requirements.txt
torch
torchvision
torchaudio
transformers
datasets
sacrebleu
sacremetrics
sentencepiece
torchtext
torchdata
numpy
pandas
scikit-learn
matplotlib
tqdm
fastapi
uvicorn
python-Levenshtein
flask
pytest
